requirements.txt
blinker==1.4
Click==7.0
Flask==1.0.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.5
PyJWT==1.6.4
SQLAlchemy==1.2.12
Werkzeug==0.14.1
